restoring old vapes for posterity and separating them from the new stuff
