+++ Content.Server/Nutrition/EntitySystems/SmokingSystem.cs
@@ -52,6 +52,7 @@ namespace Content.Server.Nutrition.EntitySystems
             InitializeCigars();
             InitializePipes();
             InitializeVapes();
+            InitializeImpVapes(); // imp
         }
 
         public void SetSmokableState(EntityUid uid, SmokableState state, SmokableComponent? smokable = null,
