+++ Resources/Locale/en-US/_Impstation/nutrition/components/vape-component.ftl
@@ -1,0 +1,6 @@
+vapepen-component-vape-success = Puff. {$flavors}
+vapepen-component-vape-success-forced = {CAPITALIZE(THE($user))} forced you to puff on the vape. {$flavors}
+vapepen-component-vape-cart-slot-empty = The vape has no cart!
+vapecart-component-vape-cart-solution-empty = The cart is empty!
+vapecart-component-vape-incompatible = The vape doesn't produce any vapor!
+vapepen-component-vape-cart-slot-name-default = Cart slot
