+++ Resources/Prototypes/Entities/Objects/Consumable/Smokeables/Vapes/vape.yml
@@ -1,173 +1,9 @@
-# IMPSTATION VAPE REWORK AHEAD! NAUGHT BUT SCRAPS OF THE ORIGINAL!
-## Vape Types ##
-
 - type: entity
-  parent: BaseVapeSwappable
-  id: VapeSwappable
+  id: Vape
+  parent: BaseVape
   name: vape
-  description: Like a cigar, but for tough teens.
+  description: "Like a cigar, but for tough teens. (WARNING:Pour only water into the vape)"
   components:
-  - type: Item
-    size: Small
-    shape:
-    - 0,0,0,1
   - type: Sprite
-    sprite: _Impstation/Objects/Consumable/Smokeables/Vapes/vape-swappable.rsi
-    layers:
-    - state: base
-      map: [ "base" ]
-    - state: cart-base
-      map: [ "enum.SolutionContainerLayers.Base" ]
-      visible: false
-    - state: fill2
-      map: [ "enum.SolutionContainerLayers.Fill" ]
-      visible: false
-    - state: cart-front
-      map: [ "enum.SolutionContainerLayers.Overlay" ]
-      visible: false
-    - state: led
-      map: [ "enum.PowerDeviceVisualLayers.Powered" ]
-      visible: false
-
-- type: entity
-  parent: BaseVapeDisposable
-  id: VapeDisposable
-  name: disposable vape
-  description: Kinda trash but gets the job done.
-  components:
-  - type: Item
-    size: Small
-    shape:
-    - 0,0,0,0
-  - type: Sprite
-    sprite: _Impstation/Objects/Consumable/Smokeables/Vapes/vape-disposable.rsi
-    layers:
-    - state: base
-      map: [ "enum.SolutionContainerLayers.Base" ]
-    - state: fill5
-      map: [ "enum.SolutionContainerLayers.Fill" ]
-      visible: false
-  - type: SolutionContainerVisuals
-    maxFillLevels: 5
-    fillBaseName: fill
-    changeColor: false
-
-
-## Cartridge Prefill Packs ##
-
-- type: entity
-  parent: BaseStorageItem
-  id: BaseVapeCartPack
-  name: vape cartridge pack
-  components:
-  - type: Appearance
-  - type: Sprite
-    sprite: _Impstation/Objects/Consumable/Smokeables/Vapes/vape-cart.rsi
-    state: box
-  - type: PhysicalComposition
-    materialComposition:
-      Plastic: 50
-  - type: Item
-    size: Small
-    shape:
-    - 0,0,1,0
-  - type: Storage
-    grid:
-    - 0,0,3,0
-  - type: StorageFill
-    contents:
-    - id: WaterVapeCart
-    - id: SugarVapeCart
-    - id: NicotineVapeCart
-      prob: 0.6
-    - id: THCVapeCart
-      prob: 0.4
-    - id: CustomVapeCart
-      prob: 0.2
-  - type: Tag
-    tags:
-    - CigPack
-    - Trash
-
-
-## Cartridge Prefills ##
-
-- type: entity
-  parent: BaseVapeCart
-  id: CustomVapeCart
-  name: DIY Cart
-  description: A vape cartridge you can fill with anything!
-  components:
-  - type: SolutionContainerManager
-    solutions:
-      smokable:
-        maxVol: 30
-  - type: DrawableSolution
-    solution: smokable
-  - type: InjectableSolution
-    solution: smokable
-  - type: VapeCart
-    flavorMultiplicationFactor: 1.0  # don't want flavor multiplication to run rampant... prefills only!
-
-- type: entity
-  parent: BaseVapeCart
-  id: SugarVapeCart
-  name: Cloud 9
-  description: A vape cartridge containing nothing but pure cotton.
-  components:
-    - type: SolutionContainerManager
-      solutions:
-        smokable:
-          maxVol: 15
-          reagents:
-          - ReagentId: Sugar
-            Quantity: 15
-
-- type: entity
-  parent: BaseVapeCart
-  id: WaterVapeCart
-  name: Vape World No Mercy
-  description: A vape cartridge containing nothing but pure water.
-  components:
-    - type: SolutionContainerManager
-      solutions:
-        smokable:
-          maxVol: 15
-          reagents:
-          - ReagentId: Water
-            Quantity: 15
-    - type: VapeCart
-      gasReductionFactor: 0.5
-      ignoreDamage: true
-
-- type: entity
-  parent: BaseVapeCart
-  id: NicotineVapeCart
-  name: Paleonicotinoids
-  description: Your new old nic fix.
-  components:
-    - type: SolutionContainerManager
-      solutions:
-        smokable:
-          maxVol: 15
-          reagents:
-          - ReagentId: Sugar
-            Quantity: 10
-          - ReagentId: Nicotine
-            Quantity: 5
-
-- type: entity
-  parent: BaseVapeCart
-  id: THCVapeCart
-  name: Cloud 10
-  description: this shit is strong bro be careful
-  components:
-    - type: SolutionContainerManager
-      solutions:
-        smokable:
-          maxVol: 15
-          reagents:
-          - ReagentId: Sugar
-            Quantity: 10
-          - ReagentId: THC
-            Quantity: 5
+    sprite: Objects/Consumable/Smokeables/Vapes/vape-standard.rsi
+    state: icon
